fix(schematics): remove local ng utils for schematics
closes #2183
