--- packages/store/schematics/src/utils/ng-utils/ng-ast-utils.ts
@@ -1,161 +1,0 @@
-import { normalize } from '@angular-devkit/core';
-import { SchematicsException, Tree } from '@angular-devkit/schematics';
-import { dirname } from 'path';
-import * as ts from '@schematics/angular/third_party/github.com/Microsoft/TypeScript/lib/typescript';
-import { findNode, getSourceNodes } from '@schematics/angular/utility/ast-utils';
-
-export function findBootstrapApplicationCall(
-  sourceFile: ts.SourceFile
-): ts.CallExpression | null {
-  const localName = findImportLocalName(
-    sourceFile,
-    'bootstrapApplication',
-    '@angular/platform-browser'
-  );
-
-  if (!localName) {
-    return null;
-  }
-
-  let result: ts.CallExpression | null = null;
-
-  sourceFile.forEachChild(function walk(node) {
-    if (
-      ts.isCallExpression(node) &&
-      ts.isIdentifier(node.expression) &&
-      node.expression.text === localName
-    ) {
-      result = node;
-    }
-
-    if (!result) {
-      node.forEachChild(walk);
-    }
-  });
-
-  return result;
-}
-
-/**
- * Finds the local name of an imported symbol. Could be the symbol name itself or its alias.
- * @param sourceFile File within which to search for the import.
- * @param name Actual name of the import, not its local alias.
- * @param moduleName Name of the module from which the symbol is imported.
- */
-function findImportLocalName(
-  sourceFile: ts.SourceFile,
-  name: string,
-  moduleName: string
-): string | null {
-  for (const node of sourceFile.statements) {
-    // Only look for top-level imports.
-    if (
-      !ts.isImportDeclaration(node) ||
-      !ts.isStringLiteral(node.moduleSpecifier) ||
-      node.moduleSpecifier.text !== moduleName
-    ) {
-      continue;
-    }
-
-    // Filter out imports that don't have the right shape.
-    if (
-      !node.importClause ||
-      !node.importClause.namedBindings ||
-      !ts.isNamedImports(node.importClause.namedBindings)
-    ) {
-      continue;
-    }
-
-    // Look through the elements of the declaration for the specific import.
-    for (const element of node.importClause.namedBindings.elements) {
-      if ((element.propertyName || element.name).text === name) {
-        // The local name is always in `name`.
-        return element.name.text;
-      }
-    }
-  }
-
-  return null;
-}
-
-export function findBootstrapModuleCall(
-  host: Tree,
-  mainPath: string
-): ts.CallExpression | null {
-  const mainText = host.readText(mainPath);
-  const source = ts.createSourceFile(mainPath, mainText, ts.ScriptTarget.Latest, true);
-
-  const allNodes = getSourceNodes(source);
-
-  let bootstrapCall: ts.CallExpression | null = null;
-
-  for (const node of allNodes) {
-    let bootstrapCallNode: ts.Node | null = null;
-    bootstrapCallNode = findNode(node, ts.SyntaxKind.Identifier, 'bootstrapModule');
-
-    // Walk up the parent until CallExpression is found.
-    while (
-      bootstrapCallNode &&
-      bootstrapCallNode.parent &&
-      bootstrapCallNode.parent.kind !== ts.SyntaxKind.CallExpression
-    ) {
-      bootstrapCallNode = bootstrapCallNode.parent;
-    }
-
-    if (
-      bootstrapCallNode !== null &&
-      bootstrapCallNode.parent !== undefined &&
-      bootstrapCallNode.parent.kind === ts.SyntaxKind.CallExpression
-    ) {
-      bootstrapCall = bootstrapCallNode.parent as ts.CallExpression;
-      break;
-    }
-  }
-
-  return bootstrapCall;
-}
-
-function findBootstrapModulePath(host: Tree, mainPath: string): string {
-  const bootstrapCall = findBootstrapModuleCall(host, mainPath);
-  if (!bootstrapCall) {
-    throw new SchematicsException('Bootstrap call not found');
-  }
-
-  const bootstrapModule = bootstrapCall.arguments[0];
-
-  const mainText = host.readText(mainPath);
-  const source = ts.createSourceFile(mainPath, mainText, ts.ScriptTarget.Latest, true);
-  const allNodes = getSourceNodes(source);
-  const bootstrapModuleRelativePath = allNodes
-    .filter(ts.isImportDeclaration)
-    .filter(imp => {
-      return findNode(imp, ts.SyntaxKind.Identifier, bootstrapModule.getText());
-    })
-    .map(imp => {
-      const modulePathStringLiteral = imp.moduleSpecifier as ts.StringLiteral;
-
-      return modulePathStringLiteral.text;
-    })[0];
-
-  return bootstrapModuleRelativePath;
-}
-
-export function getAppModulePath(host: Tree, mainPath: string): string {
-  const moduleRelativePath = findBootstrapModulePath(host, mainPath);
-  const mainDir = dirname(mainPath);
-  const modulePath = normalize(`/${mainDir}/${moduleRelativePath}.ts`);
-
-  return modulePath;
-}
-
-export function isStandaloneApp(host: Tree, mainPath: string): boolean {
-  const source = ts.createSourceFile(
-    mainPath,
-    host.readText(mainPath),
-    ts.ScriptTarget.Latest,
-    true
-  );
-  const bootstrapCall = findBootstrapApplicationCall(source);
-
-  return bootstrapCall !== null;
-}
