--- packages/store/schematics/src/utils/ng-utils/standalone/index.ts
@@ -1,1 +1,0 @@
-export { addRootImport, addRootProvider } from './rules';
